+++ src/lib.rs
@@ -2,6 +2,8 @@ pub mod ast;
 pub mod tokenizer;
 pub mod visitors;
 
+use std::fmt;
+
 #[cfg(all(test, not(feature = "serde")))]
 compile_error!("Serde feature must be enabled for tests");
 
@@ -10,6 +12,17 @@ pub enum Error<'a> {
     AstError(ast::AstError<'a>),
     TokenizerError(tokenizer::TokenizerError),
 }
+
+impl<'a> fmt::Display for Error<'a> {
+    fn fmt(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
+        match self {
+            Error::AstError(error) => write!(formatter, "error occurred while creating ast: {}", error),
+            Error::TokenizerError(error) => write!(formatter, "error occurred while tokenizing: {}", error),
+        }
+    }
+}
+
+impl<'a> std::error::Error for Error<'a> {}
 
 /// Creates an [Ast](ast/struct.Ast.html) from Lua code
 pub fn parse(code: &str) -> Result<ast::Ast, Error> {