+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added std::fmt::Display and std::error::Error implementations for Error, AstError, and TokenizerError
+
 ### Changed
 - Properties are no longer `pub` and now have public accessor methods